+++ src/test/java/io/github/astrapi69/gen/tree/visitor/ReindexTreeNodeVisitor.java
@@ -1,8 +1,8 @@
 /**
  * The MIT License
- *
+ * <p>
  * Copyright (C) 2015 Asterios Raptis
- *
+ * <p>
  * Permission is hereby granted, free of charge, to any person obtaining
  * a copy of this software and associated documentation files (the
  * "Software"), to deal in the Software without restriction, including
@@ -10,10 +10,10 @@
  * distribute, sublicense, and/or sell copies of the Software, and to
  * permit persons to whom the Software is furnished to do so, subject to
  * the following conditions:
- *
+ * <p>
  * The above copyright notice and this permission notice shall be
  * included in all copies or substantial portions of the Software.
- *
+ * <p>
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
  * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
  * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
@@ -22,34 +22,42 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package io.github.astrapi69.gen.tree.api;
+package io.github.astrapi69.gen.tree.visitor;
 
-import io.github.astrapi69.data.identifiable.GenericIdentifiable;
-import io.github.astrapi69.gen.tree.handler.IBaseTreeNodeHandlerExtensions;
-import lombok.NonNull;
+import io.github.astrapi69.data.identifiable.IdGenerator;
+import io.github.astrapi69.design.pattern.visitor.Visitor;
+import io.github.astrapi69.gen.tree.api.IBaseTreeNode;
+import io.github.astrapi69.gen.tree.api.ITreeNode;
+import lombok.Getter;
+
+import java.util.Collection;
 
 /**
- * The Interface {@link IBaseTreeNode} extends {@link ITreeNode} and provides an additional id field
- * that can be used as key
+ * This visitor visits all {@link ITreeNode} objects and adds them to a {@link Collection} object
+ * with all descendant
  *
- * @param <V>
- *            the generic type of the value
- * @param <K>
- *            the generic type of the id of the node
  * @param <T>
- *            the generic type of the concrete tree node
+ *            the generic type of the value
  */
-public interface IBaseTreeNode<V, K, T extends IBaseTreeNode<V, K, T>>
-	extends
-		ITreeNode<V, T>,
-		GenericIdentifiable<K>
+public class ReindexTreeNodeVisitor<V, K, T extends IBaseTreeNode<V, K, T>> implements Visitor<T>
 {
+	/**
+	 * The {@link IdGenerator} object for reindex tree ids
+	 */
+	@Getter
+	private final IdGenerator<K> idGenerator;
+
+	public ReindexTreeNodeVisitor(final IdGenerator<K> idGenerator)
+	{
+		this.idGenerator = idGenerator;
+	}
 
 	/**
 	 * {@inheritDoc}
 	 */
-	default T findById(final @NonNull K id)
+	@Override
+	public void visit(T treeNode)
 	{
-		return IBaseTreeNodeHandlerExtensions.findById((T)this, id);
+		treeNode.setId(idGenerator.getNextId());
 	}
 }